build.mk, params.txt: add parameter USE_QEMU
The parameter determines use of QEMU or not, if the target architecture
does not correspond to the host architecture.
By default, the parameter is on (Value 1).
For architectures that do not support QEMU (e2k), the option is turned off.

(cherry picked from commit 1c50ff1ffa9d6460348d7ccf4172313d41851c8a)