create initrd.img with propagator instead full.cz
With this change, we simplify the created initrd image. Kernel
modules and firmware are added by make-Initrd itself. The layer
in Initrd is now one. It also allows you to unifyed configs for
bootloading with propagator and bootchain.
Config files for refind and elilo is creates in mkimage, so have
to copy initrd.img in efi/boot/ with the name full.cz.

(cherry picked from commit 71e209e2013a553016281ea949990d02e20a923f)